modularizando funciones de eventos
